+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/case-editor/case-edit/case-edit.component.spec.ts
@@ -4,23 +4,23 @@ import { FormArray, FormControl, FormGroup, ReactiveFormsModule } from '@angular
 import { ActivatedRoute, Router } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { MockComponent } from 'ng2-mock-component';
-import { Observable, of } from 'rxjs';
+import { Observable, of, throwError } from 'rxjs';
 import { ConditionalShowRegistrarService } from '../../../directives';
-import { FieldType, Profile } from '../../../domain';
+import { CaseView, FieldType, HttpError, Profile } from '../../../domain';
 import { CaseEventTrigger } from '../../../domain/case-view/case-event-trigger.model';
 import { CaseField } from '../../../domain/definition/case-field.model';
 import { createCaseEventTrigger } from '../../../fixture/shared.test.fixture';
 import { FieldsFilterPipe } from '../../../pipes/complex/fields-filter.pipe';
-import { FieldsPurger, FieldsUtils, SessionStorageService, WindowService } from '../../../services';
+import { FieldsPurger, FieldsUtils, LoadingService, SessionStorageService, WindowService } from '../../../services';
 import { FormErrorService, FormValueService } from '../../../services/form';
 import { PaletteUtilsModule } from '../../palette';
-import { Wizard, WizardPage, WizardPageField } from '../domain';
+import { Confirmation, Wizard, WizardPage, WizardPageField } from '../domain';
+import { CaseNotifier } from '../services';
 import { WizardFactoryService } from '../services/wizard-factory.service';
 import { CaseEditComponent } from './case-edit.component';
 import createSpyObj = jasmine.createSpyObj;
 
 describe('CaseEditComponent', () => {
-
   const EVENT_TRIGGER: CaseEventTrigger = createCaseEventTrigger(
     'TEST_TRIGGER',
     'Test Trigger',
@@ -179,22 +179,22 @@ describe('CaseEditComponent', () => {
   let component: CaseEditComponent;
   let de: DebugElement;
 
-  const EventTriggerHeaderComponent: any = MockComponent({
+  const eventTriggerHeaderComponentMock: any = MockComponent({
     selector: 'ccd-event-trigger-header',
     inputs: ['eventTrigger']
   });
 
-  const FieldRead: any = MockComponent({
+  const fieldReadComponentMock: any = MockComponent({
     selector: 'ccd-field-read',
     inputs: ['caseField']
   });
 
-  const FieldWrite: any = MockComponent({
+  const fieldWriteComponentMock: any = MockComponent({
     selector: 'ccd-field-write',
     inputs: ['caseField', 'formGroup', 'idPrefix', 'isExpanded', 'parent']
   });
 
-  const RouterLinkComponent: any = MockComponent({
+  const routerLinkComponentMock: any = MockComponent({
     selector: 'a',
     inputs: ['routerLink']
   });
@@ -243,6 +243,7 @@ describe('CaseEditComponent', () => {
       isSolicitor: FUNC,
       isCourtAdmin: FUNC
     };
+    let loadingServiceMock: jasmine.SpyObj<LoadingService>;
 
     beforeEach(waitForAsync(() => {
       cancelHandler = createSpyObj('cancelHandler', ['applyFilters']);
@@ -256,7 +257,17 @@ describe('CaseEditComponent', () => {
       wizard.pages = [];
       formErrorService = createSpyObj<FormErrorService>('formErrorService', ['mapFieldErrors']);
 
-      formValueService = createSpyObj<FormValueService>('formValueService', ['sanitise']);
+      formValueService = createSpyObj<FormValueService>('formValueService', [
+        'sanitise',
+        'clearNonCaseFields',
+        'removeNullLabels',
+        'removeEmptyDocuments',
+        'populateFlagDetailsFromCaseFields',
+        'removeFlagLauncherField',
+        'removeComponentLauncherField',
+        'removeEmptyCollectionsWithMinValidation',
+        'populateLinkedCasesDetailsFromCaseFields'
+      ]);
       mockSessionStorageService = createSpyObj<SessionStorageService>('SessionStorageService', ['getItem', 'removeItem', 'setItem']);
 
       route = {
@@ -276,6 +287,8 @@ describe('CaseEditComponent', () => {
         params: of({})
       };
 
+      loadingServiceMock = createSpyObj<LoadingService>('loadingService', ['register', 'unregister']);
+
       TestBed
         .configureTestingModule({
           imports: [
@@ -286,15 +299,16 @@ describe('CaseEditComponent', () => {
           declarations: [
             CaseEditComponent,
 
-            // Mock
-            EventTriggerHeaderComponent,
-            RouterLinkComponent,
+            // Mocks
+            eventTriggerHeaderComponentMock,
+            routerLinkComponentMock,
             FieldsFilterPipe,
-            FieldRead,
-            FieldWrite
+            fieldReadComponentMock,
+            fieldWriteComponentMock
           ],
           providers: [
             WizardFactoryService,
+            {provide: CaseNotifier, useValue: { cachedCaseView: null}},
             {provide: FormErrorService, useValue: formErrorService},
             {provide: FormValueService, useValue: formValueService},
             {provide: FieldsUtils, useValue: fieldsUtils},
@@ -303,7 +317,8 @@ describe('CaseEditComponent', () => {
             {provide: Router, useValue: routerStub},
             {provide: ActivatedRoute, useValue: route},
             SessionStorageService,
-            WindowService
+            WindowService,
+            { provide: LoadingService, loadingServiceMock }
           ]
         })
         .compileComponents();
@@ -318,10 +333,6 @@ describe('CaseEditComponent', () => {
       de = fixture.debugElement;
       fixture.detectChanges();
     }));
-
-    beforeEach(() => {
-    });
-
     // Moved this test case to case-edit-page.component
 
     it('should return true for hasPrevious', () => {
@@ -342,9 +353,7 @@ describe('CaseEditComponent', () => {
     });
 
     describe('fieldShowCondition', () => {
-
       describe('next page', () => {
-
         it('should navigate to next page when next is called and do not clear READONLY hidden field value', () => {
           component.wizard = wizard;
           const currentPage = new WizardPage();
@@ -580,9 +589,7 @@ describe('CaseEditComponent', () => {
     });
 
     describe('pageShowCondition', () => {
-
       describe('next page', () => {
-
         it('should navigate to next page when next is called and do not clear visible field', () => {
           component.wizard = wizard;
           const currentPage = new WizardPage();
@@ -806,7 +813,6 @@ describe('CaseEditComponent', () => {
       });
 
       describe('previous page', () => {
-
         it('should navigate to previous page when previous is called and do not clear visible field', () => {
           component.wizard = wizard;
           const currentPage = new WizardPage();
@@ -837,8 +843,7 @@ describe('CaseEditComponent', () => {
           expect(component.form.get('data').get(CASE_FIELD_3.id)).not.toBeNull();
         });
 
-        it('should navigate to previous page when previous is called and retain hidden simple form fields with ' +
-        'retain_hidden_value = true', () => {
+        it('should navigate to previous page when previous is called and retain hidden simple form fields with retain_hidden_value = true', () => {
           component.wizard = wizard;
           const currentPage = new WizardPage();
           currentPage.wizard_page_fields = [WIZARD_PAGE_1];
@@ -1028,10 +1033,219 @@ describe('CaseEditComponent', () => {
       component.cancel();
       expect(cancelHandler.applyFilters).toHaveBeenCalled();
     });
+
+    describe('getStatus', () => {
+      it('should get callback_response_status as vallback has failed', () => {
+        const response = {
+          /* tslint:disable:object-literal-key-quotes */
+          'delete_draft_response_status': 'delete_draft_response_status',
+          /* tslint:disable:object-literal-key-quotes */
+          'callback_response_status': 'CALLBACK_HASNOT_COMPLETED'
+        }
+        const actual = component.getStatus(response);
+        expect(actual).toEqual(response['callback_response_status']);
+      });
+
+      it('should get callback_response_status', () => {
+        const response = {
+          /* tslint:disable:object-literal-key-quotes */
+          'delete_draft_response_status': 'delete_draft_response_status',
+          /* tslint:disable:object-literal-key-quotes */
+          'callback_response_status': 'CALLBACK_COMPLETED'
+        }
+        const actual = component.getStatus(response);
+        expect(actual).toEqual(response['delete_draft_response_status']);
+      });
+    })
+
+    describe('emitSubmitted', () => {
+      it('should emit submitted', () => {
+        const response = {
+          /* tslint:disable:object-literal-key-quotes */
+          'delete_draft_response_status': 'delete_draft_response_status',
+          /* tslint:disable:object-literal-key-quotes */
+          'callback_response_status': 'CALLBACK_HASNOT_COMPLETED'
+        }
+        spyOn(component.submitted, 'emit');
+        component.emitSubmitted(response);
+
+        expect(component.submitted.emit).toHaveBeenCalled();
+      });
+    })
+
+    describe('confirm', () => {
+      it('should call routers navigate', () => {
+        component.confirm({} as unknown as Confirmation);
+
+        expect(routerStub.navigate).toHaveBeenCalled();
+        expect(component.confirmation).toEqual({} as unknown as Confirmation);
+      });
+    })
+
+    describe('getCaseId', () => {
+      it('should return case id', () => {
+        const actual = component.getCaseId({
+          /* tslint:disable:object-literal-key-quotes */
+          'case_id': 'case_id'
+        } as unknown as CaseView);
+
+        expect(actual).toEqual('case_id');
+      });
+    })
+
+    describe('submitForm', () => {
+      it('should submit case', () => {
+        const mockClass = {
+          submit: () => of({})
+        }
+        spyOn(mockClass, 'submit').and.returnValue(of({
+          id: 'id',
+          /* tslint:disable:object-literal-key-quotes */
+          'callback_response_status': 'CALLBACK_HASNOT_COMPLETED',
+          /* tslint:disable:object-literal-key-quotes */
+          'after_submit_callback_response': {
+          /* tslint:disable:object-literal-key-quotes */
+            'confirmation_header': 'confirmation_header',
+          /* tslint:disable:object-literal-key-quotes */
+            'confirmation_body': 'confirmation_body'
+          }
+        }))
+        formValueService.sanitise.and.returnValue({name: 'sweet'})
+
+        fixture.detectChanges();
+
+        component.submitForm({
+          eventTrigger: component.eventTrigger,
+          caseDetails: component.caseDetails,
+          form: component.form,
+          submit: mockClass.submit,
+        });
+
+        expect(component.isSubmitting).toEqual(true);
+        expect(formValueService.sanitise).toHaveBeenCalled();
+      });
+    })
+
+    describe('onEventCanBeCompleted', () => {
+      it('should submit the case', () => {
+        const mockClass = {
+          submit: () => of({})
+        }
+        spyOn(mockClass, 'submit').and.returnValue(of({
+          id: 'id',
+          /* tslint:disable:object-literal-key-quotes */
+          'callback_response_status': 'CALLBACK_HASNOT_COMPLETED',
+          /* tslint:disable:object-literal-key-quotes */
+          'after_submit_callback_response': {
+          /* tslint:disable:object-literal-key-quotes */
+            'confirmation_header': 'confirmation_header',
+          /* tslint:disable:object-literal-key-quotes */
+            'confirmation_body': 'confirmation_body'
+          }
+        }))
+
+        spyOn(component, 'confirm')
+
+        component.isCaseFlagSubmission = true;
+        component.isLinkedCasesSubmission = true;
+        component.confirmation = {} as unknown as Confirmation;
+
+        formValueService.sanitise.and.returnValue({name: 'sweet'})
+        component.onEventCanBeCompleted({
+          eventTrigger: component.eventTrigger,
+          eventCanBeCompleted: true,
+          caseDetails: component.caseDetails,
+          form: component.form,
+          submit: mockClass.submit,
+        });
+
+        expect(component.confirm).toHaveBeenCalled();
+        expect(formValueService.populateLinkedCasesDetailsFromCaseFields).toHaveBeenCalled();
+        expect(formValueService.removeComponentLauncherField).toHaveBeenCalled();
+        expect(formValueService.removeFlagLauncherField).toHaveBeenCalled();
+        expect(formValueService.populateFlagDetailsFromCaseFields).toHaveBeenCalled();
+      });
+
+      it('should NOT submit the case due to error', () => {
+        const mockClass = {
+          submit: () => of({})
+        }
+        spyOn(mockClass, 'submit').and.returnValue(throwError(({ details: 'details' } as unknown as HttpError)))
+        spyOn(component, 'confirm')
+
+        component.isCaseFlagSubmission = true;
+        component.isLinkedCasesSubmission = true;
+        component.confirmation = {} as unknown as Confirmation;
+
+        formValueService.sanitise.and.returnValue({name: 'sweet'})
+        component.onEventCanBeCompleted({
+          eventTrigger: component.eventTrigger,
+          eventCanBeCompleted: true,
+          caseDetails: component.caseDetails,
+          form: component.form,
+          submit: mockClass.submit,
+        });
+
+        expect(formErrorService.mapFieldErrors).toHaveBeenCalled();
+        expect(component.error).toEqual({ details: 'details' } as unknown as HttpError);
+        expect(component.isSubmitting).toEqual(false);
+      });
+
+      it('should NOT submit the case and should navigate to tasks tab', () => {
+        const mockClass = {
+          submit: () => of({})
+        }
+        spyOn(mockClass, 'submit');
+
+        spyOn(component, 'confirm')
+
+        component.isCaseFlagSubmission = true;
+        component.isLinkedCasesSubmission = true;
+        component.confirmation = {} as unknown as Confirmation;
+
+        formValueService.sanitise.and.returnValue({name: 'sweet'})
+        component.onEventCanBeCompleted({
+          eventTrigger: component.eventTrigger,
+          eventCanBeCompleted: false,
+          caseDetails: component.caseDetails,
+          form: component.form,
+          submit: mockClass.submit,
+        });
+
+        expect(component.confirm).not.toHaveBeenCalled();
+        expect(mockClass.submit).not.toHaveBeenCalled();
+        expect(routerStub.navigate).toHaveBeenCalled();
+      });
+
+      it('should emit submit as after_submit_callback_response is NOT present in response', () => {
+        const mockClass = {
+          submit: () => of({})
+        }
+        spyOn(mockClass, 'submit').and.returnValue(of({
+          id: 'id',
+          /* tslint:disable:object-literal-key-quotes */
+          'callback_response_status': 'CALLBACK_HASNOT_COMPLETED'
+        }))
+
+        spyOn(component, 'confirm');
+        spyOn(component, 'emitSubmitted');
+
+        formValueService.sanitise.and.returnValue({name: 'sweet'})
+        component.onEventCanBeCompleted({
+          eventTrigger: component.eventTrigger,
+          eventCanBeCompleted: true,
+          caseDetails: component.caseDetails,
+          form: component.form,
+          submit: mockClass.submit,
+        });
+
+        expect(component.confirm).not.toHaveBeenCalled();
+        expect(component.emitSubmitted).toHaveBeenCalled();
+      });
+    })
   });
 
   xdescribe('profile not available in route', () => {
-
     const USER = {
       idam: {
         id: 'userId',
@@ -1110,11 +1324,11 @@ describe('CaseEditComponent', () => {
           declarations: [
             CaseEditComponent,
 
-            // Mock
-            EventTriggerHeaderComponent,
-            RouterLinkComponent,
-            FieldRead,
-            FieldWrite
+            // Mocks
+            eventTriggerHeaderComponentMock,
+            routerLinkComponentMock,
+            fieldReadComponentMock,
+            fieldWriteComponentMock
           ],
           providers: [
             WizardFactoryService,