+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/error/callback-errors.component.spec.ts
@@ -62,8 +62,8 @@ describe('CallbackErrorsComponent', () => {
     callbackErrorsSubject.next(error);
 
     const expectedCallbackErrorsContext: CallbackErrorsContext = new CallbackErrorsContext();
-    expectedCallbackErrorsContext.ignore_warning = true;
-    expectedCallbackErrorsContext.trigger_text = triggerTextIgnore;
+    expectedCallbackErrorsContext.ignoreWarning = true;
+    expectedCallbackErrorsContext.triggerText = triggerTextIgnore;
     expect(mockCallbackErrorsContext.emit).toHaveBeenCalledWith(expectedCallbackErrorsContext);
   });
 
@@ -73,8 +73,8 @@ describe('CallbackErrorsComponent', () => {
     callbackErrorsSubject.next(error);
 
     const expectedCallbackErrorsContext: CallbackErrorsContext = new CallbackErrorsContext();
-    expectedCallbackErrorsContext.ignore_warning = false;
-    expectedCallbackErrorsContext.trigger_text = triggerTextContinue;
+    expectedCallbackErrorsContext.ignoreWarning = false;
+    expectedCallbackErrorsContext.triggerText = triggerTextContinue;
     expect(mockCallbackErrorsContext.emit).toHaveBeenCalledWith(expectedCallbackErrorsContext);
   });
 
@@ -83,8 +83,8 @@ describe('CallbackErrorsComponent', () => {
     callbackErrorsSubject.next(error);
 
     const expectedCallbackErrorsContext: CallbackErrorsContext = new CallbackErrorsContext();
-    expectedCallbackErrorsContext.ignore_warning = false;
-    expectedCallbackErrorsContext.trigger_text = triggerTextContinue;
+    expectedCallbackErrorsContext.ignoreWarning = false;
+    expectedCallbackErrorsContext.triggerText = triggerTextContinue;
     expect(mockCallbackErrorsContext.emit).toHaveBeenCalledWith(expectedCallbackErrorsContext);
   });
 
@@ -130,8 +130,8 @@ describe('CallbackErrorsComponent', () => {
 
     expect(component.error).toEqual(httpError);
     const expectedCallbackErrorsContext: CallbackErrorsContext = new CallbackErrorsContext();
-    expectedCallbackErrorsContext.ignore_warning = false;
-    expectedCallbackErrorsContext.trigger_text = triggerTextContinue;
+    expectedCallbackErrorsContext.ignoreWarning = false;
+    expectedCallbackErrorsContext.triggerText = triggerTextContinue;
     expect(mockCallbackErrorsContext.emit).toHaveBeenCalledWith(expectedCallbackErrorsContext);
   });