+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/components/manage-case-flags/manage-case-flags.component.spec.ts
@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { CaseField } from '../../../../../domain';
 import { FlagDetail, FlagDetailDisplayWithFormGroupPath, FlagsWithFormGroupPath } from '../../domain';
@@ -117,7 +117,7 @@ describe('ManageCaseFlagsComponent', () => {
     }
   ] as FlagsWithFormGroupPath[];
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       imports: [ ReactiveFormsModule ],
       schemas: [ CUSTOM_ELEMENTS_SCHEMA ],
@@ -255,7 +255,7 @@ describe('ManageCaseFlagsComponent', () => {
         partyName: 'Wayne Sleep'
       },
       pathToFlagsFormGroup: 'CaseFlag2'
-    }
+    };
 
     const displayResult = component.mapFlagDetailForDisplay(flagDetail, flagsInstance);
     expect(displayResult.flagDetailDisplay.partyName).toEqual(flagsInstance.flags.partyName);
@@ -341,7 +341,7 @@ describe('ManageCaseFlagsComponent', () => {
       },
       pathToFlagsFormGroup: 'caseFlags',
       caseField: flagsData[2].caseField
-    } as FlagDetailDisplayWithFormGroupPath
+    } as FlagDetailDisplayWithFormGroupPath;
     expect(component.getPartyName(flagDisplay)).toEqual('Case level');
     flagDisplay.pathToFlagsFormGroup = null;
     expect(component.getPartyName(flagDisplay)).toEqual('');
@@ -366,7 +366,7 @@ describe('ManageCaseFlagsComponent', () => {
       pathToFlagsFormGroup: '',
       caseField: flagsData[2].caseField,
       roleOnCase: 'Applicant'
-    } as FlagDetailDisplayWithFormGroupPath
+    } as FlagDetailDisplayWithFormGroupPath;
     expect(component.getRoleOnCase(flagDisplay)).toEqual(' (Applicant)');
   });