+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/linked-cases/components/linked-cases-table/linked-cases-from-table.component.ts
@@ -15,7 +15,6 @@ import { LinkedCasesService } from '../../services';
 })
 
 export class LinkedCasesFromTableComponent implements OnInit, AfterViewInit {
-
   private static readonly CASE_NAME_MISSING_TEXT = 'Case name missing';
 
   @Input()