+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/utils/first-error.pipe.spec.ts
@@ -1,9 +1,8 @@
 import { FirstErrorPipe } from './first-error.pipe';
 
 describe('FirstErrorPipe', () => {
-
   const ERROR_MESSAGE = 'This is wrong';
-  const Field_Label = 'Field1';
+  const FIELD_LABEL = 'Field1';
 
   let firstError: FirstErrorPipe;
 
@@ -43,32 +42,32 @@ describe('FirstErrorPipe', () => {
   it('should return exact error along with label name when field value is MANDATORY', () => {
     const message = firstError.transform({
       required: true
-    }, Field_Label);
+    }, FIELD_LABEL);
 
-    expect(message).toBe(`${Field_Label} is required`);
+    expect(message).toBe(`${FIELD_LABEL} is required`);
   });
 
   it('should return exact error along with label name when pattern does not match', () => {
     const message = firstError.transform({
       pattern: {actualValue: 'test ', requiredPattern: '^[0-9 +().-]{9,}$'}
-    }, Field_Label);
+    }, FIELD_LABEL);
 
-    expect(message).toBe(`The data entered is not valid for ${Field_Label}`);
+    expect(message).toBe(`The data entered is not valid for ${FIELD_LABEL}`);
   });
 
   it('should return exact error along with label name when field value is below minimum length', () => {
     const message = firstError.transform({
       minlength: {actualValue: 'test', requiredLength: 5}
-    }, Field_Label);
+    }, FIELD_LABEL);
 
-    expect(message).toBe(`${Field_Label} is below the minimum length`);
+    expect(message).toBe(`${FIELD_LABEL} is below the minimum length`);
   });
 
   it('should return exact error along with label name when field value exceeds maximum length', () => {
     const message = firstError.transform({
       maxlength: {actualValue: 'test is done', requiredLength: 10}
-    }, Field_Label);
+    }, FIELD_LABEL);
 
-    expect(message).toBe(`${Field_Label} exceeds the maximum length`);
+    expect(message).toBe(`${FIELD_LABEL} exceeds the maximum length`);
   });
 });