+++ projects/ccd-case-ui-toolkit/src/lib/shared/pipes/index.ts
@@ -1,7 +1,6 @@
 export * from './case-reference';
 export * from './case-title';
 export * from './complex';
-export * from './link-cases-from-reason-code';
 export * from './link-cases-reason-code';
 export * from './link-cases-from-reason-code';
 export * from './pipes.module';