+++ RELEASE-NOTES.md
@@ -1,4 +1,12 @@
 ## RELEASE NOTES
+### Version 6.13.10-case-flags-show-language-for-interpreter-flag-types
+**EUI-8069** Fix Case Flags table display to show selected language for "Language Interpreter" flag types
+
+### Version 6.13.11-judicial-user-typeahead-fix
+**EUI-7733** Judicial user typeahead fix
+### Version 6.13.10-linked-cases-continue-button-validation-v6
+**EUI-8025** Linked cases continue button validation
+
 ### Version 6.13.10-rc1
 **EUI-5298** Release 6.13.10 including 6.13.0, linked cases (EUI-7676) and case file view fix (EUI-7417)
 
@@ -12,6 +20,10 @@
 **EUI-7246** Linked cases other description
 
 ### Version 6.10.10-linked-cases-and-case-file-view
+### Version 6.10.11-judicial-user-typeahead-fix
+**EUI-7733** Judicial user typeahead fix
+
+### Version 6.10.10-configure-linked-cases
 **EUI-7676** Configure linked cases
 **EUI-7417** Set boolean flag allowing Case File View document move functionality in readonly mode to `false`
 
@@ -31,9 +43,6 @@
 
 ### Version 5.0.50-sscs-sit-release
 **EUI-7564** SSCS SIT release
-
-### Version 5.0.46-fix-case-link
-**EUI-7521** Fix case link hyperlink on applications tab
 
 ### Version 5.0.46-fix-case-link
 **EUI-7521** Fix case link hyperlink on applications tab